requirements.txt
APScheduler==3.7.0
asgiref==3.6.0
certifi==2022.12.7
chardet==4.0.0
Django==3.2.5
idna==2.10
python-decouple==3.8
pytz==2023.2
requests==2.25.1
six==1.16.0
sqlparse==0.4.3
tzlocal==2.1
urllib3==1.26.15
